+++ src/app/constants/index.ts
@@ -1,5 +1,6 @@
 export * from "./error-messages.constants";
 export * from "./environment.constants";
+export * from "./fail-open.constants";
 export * from "./rate-limit.constants";
 export * from "./react-translation.constants";
 export * from "./runtime.constants";